feat(sync): prompt for firmware and tracks the moment a logger connects
The firmware dialog used to be inline in FirmwareUpdateSection, which mounts
only when the drawer is open, on the Device tab, on the settings sub-tab, while
connected over BLE. A check that fires on connect had nowhere to show itself.
Split the dialog out and give a FirmwareUpdateProvider ownership of the single
useFirmwareUpdate instance, rendered once high in the tree. Mounting a second
instance instead would mean two device-version reads over GATT and two dialogs
racing to be the one on screen.

DeviceConnectFlow then runs the two questions in order on each new connection:

- FIRMWARE FIRST, because accepting it reboots the device and drops the link —
  anything queued behind it would be thrown away. It is also the rarer of the
  two, and "remind me tomorrow" keeps it rare.
- THEN TRACKS, and only when the sync plan actually has rows. A prompt that
  appears on every connect to say "nothing to sync" is worse than silence, and
  the plan already excludes the rows that could never converge.

Declining the track prompt is scoped to the connection: no stored suppression,
and the flow only re-runs when a new connection appears. Being asked again after
deliberately reconnecting is expected; being asked twice on one connection is
nagging.

Offline, the firmware step is skipped rather than surfaced — the manifest is a
network fetch and that is not this connection's problem. A failure reading the
card is logged, not toasted, for the same reason the auto-check went silent: the
user didn't ask, and the Device → Tracks tab is still there.
